+++ packages/demo/src/app/pages/movies/movies-page.component.ts
@@ -1,4 +1,4 @@
-import { Apollo, gql, onlyComplete } from 'apollo-angular';
+import { Apollo, gql, onlyCompleteData } from 'apollo-angular';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { AsyncPipe } from '@angular/common';
@@ -59,7 +59,7 @@ export class MoviesPageComponent implements OnInit {
         notifyOnNetworkStatusChange: false,
       })
       .valueChanges.pipe(
-        onlyComplete(),
+        onlyCompleteData(),
         map(result => result.data.allFilms.films),
       );
   }
